+++ src/Resources/contao/languages/de/modules.php
@@ -10,7 +10,7 @@
  * Backend-Modul Übersetzungen
  */
 $GLOBALS['TL_LANG']['MOD']['internetschach'] = array('Internetschach', 'Internetschachturniere verwalten');
-$GLOBALS['TL_LANG']['MOD']['internetschach_import_headline'] = array('Spielerdaten aus in einer ZIP-Datei gepackten CSV-Dateien importieren', 'Spielerdaten aus in einer ZIP-Datei gepackten CSV-Dateien importieren');
+$GLOBALS['TL_LANG']['MOD']['internetschach_import_headline'] = array('Spielerdaten aus CSV-Dateien importieren', 'Spielerdaten aus in einer ZIP-Datei gepackten CSV-Dateien importieren');
 $GLOBALS['TL_LANG']['MOD']['internetschach_import_file'] = array('ZIP-Datei', 'ZIP-Datei auswählen');
 $GLOBALS['TL_LANG']['MOD']['internetschach_import_submit'] = array('Import starten', 'Import starten');